requirements.txt
dill
matplotlib
numpy
pandas
Pillow
streamlit
tensorflow-cpu
